Making files ARC and 64-bit compliant.

+++ DDComet/DDCometSubscription.m
@@ -98,7 +98,7 @@
 
 -(NSString*)description
 {
-    return [NSString stringWithFormat:@"%@-%d", self.channel, self.hash];
+    return [NSString stringWithFormat:@"%@-%lu", self.channel, (unsigned long)self.hash];
 }
 
 @end
